Rename Behaviors namespace to PipelineBehaviors
